fix: break circular dependency in EventRow components causing Vite HMR loop
Extract shared types and styled components from EventRowDynamicComponent
into EventRowBase to eliminate circular imports between the parent
component and its child row variants (Message, Activity, MainObject,
CalendarEvent). This was causing Linaria/wyw to repeatedly re-evaluate
all files in the cycle during HMR.
