+++ packages/twenty-front/src/modules/activities/timeline-activities/rows/components/EventRowDynamicComponent.tsx
@@ -1,38 +1,16 @@
-import { styled } from '@linaria/react';
-
 import { EventRowActivity } from '@/activities/timeline-activities/rows/activity/components/EventRowActivity';
 import { EventRowCalendarEvent } from '@/activities/timeline-activities/rows/calendar/components/EventRowCalendarEvent';
+import {
+  type EventRowDynamicComponentProps,
+  StyledEventRowItemAction,
+  StyledEventRowItemColumn,
+} from '@/activities/timeline-activities/rows/components/EventRowBase';
 import { EventRowMainObject } from '@/activities/timeline-activities/rows/main-object/components/EventRowMainObject';
 import { EventRowMessage } from '@/activities/timeline-activities/rows/message/components/EventRowMessage';
-import { type TimelineActivity } from '@/activities/timeline-activities/types/TimelineActivity';
 import { CoreObjectNameSingular } from 'twenty-shared/types';
-import { type EnrichedObjectMetadataItem } from '@/object-metadata/types/EnrichedObjectMetadataItem';
-import { themeCssVariables } from 'twenty-ui/theme-constants';
 
-export interface EventRowDynamicComponentProps {
-  labelIdentifierValue: string;
-  event: TimelineActivity;
-  mainObjectMetadataItem: EnrichedObjectMetadataItem;
-  linkedObjectMetadataItem: EnrichedObjectMetadataItem | null;
-  authorFullName: string;
-  createdAt?: string;
-}
-
-export const StyledEventRowItemColumn = styled.div`
-  align-items: center;
-  color: ${themeCssVariables.font.color.primary};
-  display: flex;
-  flex-direction: row;
-  gap: ${themeCssVariables.spacing[1]};
-`;
-
-export const StyledEventRowItemAction = styled.div`
-  align-items: center;
-  color: ${themeCssVariables.font.color.secondary};
-  display: flex;
-  flex-direction: row;
-  gap: ${themeCssVariables.spacing[1]};
-`;
+export type { EventRowDynamicComponentProps };
+export { StyledEventRowItemAction, StyledEventRowItemColumn };
 
 export const EventRowDynamicComponent = ({
   labelIdentifierValue,
